pyproject.toml
[project]
name = "smartscreen-driver"
version = "0.2.0"
description = "Driver for serial-over-USB displays; library extracted from turing-smart-screen-python"
authors = [
  { name="Mathieu Houdebine" },
  { name="Hugo Chargois", email="hugo.chargois@free.fr" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.2",
    "pillow>=11.1.0",
    "pyserial>=3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/smartscreen_driver"]
